phase-3/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2024.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-jose[cryptography]>=3.3.0
bcrypt==4.1.3
email-validator>=2.0.0
cachetools>=5.3.0

# Phase 3: AI Chatbot dependencies
openai-agents>=0.6.4
mcp>=1.25.0
openai>=2.0.0
chatkit>=0.1.0  # OpenAI ChatKit Python SDK

# Additional dependencies from pyproject.toml
asyncpg>=0.31.0
better-auth>=0.0.1b11
openai-chatkit>=0.1.0
cryptography>=46.0.3
fastapi>=0.124.0
fastapi-cors>=0.0.6
fastmcp
pyjwt>=2.10.1
pytest-cov>=7.0.0
pydantic-settings>=2.12.0
sqlmodel>=0.0.27
starlette>=0.50.0
passlib[bcrypt]>=1.7.4
